move _set_shuffle
